IOS-2739: Refactoring WC
