+++ Tangem/App/Services/WalletConnect/WalletConnectService/Handlers/SignHandlers/BnbSignHandler.swift
