+++ Tangem/App/Services/WalletConnect/WalletConnectService/Handlers/SignHandlers/TypedData/ABIEncoder.swift
