+++ Tangem/App/Services/WalletConnect/WalletConnectService/Handlers/SignHandlers/TypedData/ABIError.swift
