+++ Tangem/App/Services/WalletConnect/WalletConnectService/Handlers/SignHandlers/TypedData/EthereumAddress.swift
