+++ Tangem/App/Services/WalletConnect/WalletConnectService/WalletConnectNetworkParserUtility.swift
